Optimize PoW
