+++ shared/shared_test.go
@@ -1,10 +1,8 @@
 package shared
 
 import (
-	"context"
 	"encoding/binary"
 	"encoding/hex"
-	"fmt"
 	"math"
 	"testing"
 
@@ -109,44 +107,5 @@ func FuzzMerkleProofSafety(f *testing.F) {
 func BenchmarkFiatShamir(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		FiatShamir([]byte("challenge this"), uint64(30000000000), T)
-	}
-}
-
-func TestCheckLeadingZeroBits(t *testing.T) {
-	r := require.New(t)
-
-	r.True(CheckLeadingZeroBits([]byte{0x00}, 0))
-	r.True(CheckLeadingZeroBits([]byte{0x00}, 8))
-
-	// Out of bounds
-	r.False(CheckLeadingZeroBits([]byte{0x00}, 9))
-
-	r.True(CheckLeadingZeroBits([]byte{0x0F}, 4))
-	r.False(CheckLeadingZeroBits([]byte{0x0F}, 5))
-
-	r.True(CheckLeadingZeroBits([]byte{0x00, 0x0F}, 5))
-	r.True(CheckLeadingZeroBits([]byte{0x00, 0x0F}, 12))
-	r.False(CheckLeadingZeroBits([]byte{0x00, 0x0F}, 13))
-}
-
-func BenchmarkFindSubmitPowNonce(b *testing.B) {
-	powChallenge := make([]byte, 32)
-	poetChallenge := make([]byte, 32)
-	nodeID := make([]byte, 32)
-	b.ResetTimer()
-	for difficulty := 0; difficulty <= 24; difficulty += 4 {
-		b.Run(fmt.Sprintf("difficulty=%v", difficulty), func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
-				binary.LittleEndian.PutUint64(nodeID, uint64(i))
-				_, err := FindSubmitPowNonce(
-					context.Background(),
-					powChallenge,
-					poetChallenge,
-					nodeID,
-					uint(difficulty),
-				)
-				require.NoError(b, err)
-			}
-		})
 	}
 }
